// src/modules/invoice/usecase/generate/generate-invoice.dto.ts
// DTO Generate
export interface GenerateInvoiceUseCaseInputDto {
	name: string;
	document: string;
	street: string;
	number: string;
	complement: string;
	city: string;
	state: string;
	zipCode: string;
	items: {
		id: string;
		name: string;
		price: number;
	}[];
}

export interface GenerateInvoiceUseCaseOutputDto {
	id: string;
	name: string;
	document: string;
	street: string;
	number: string;
	complement: string;
	city: string;
	state: string;
	zipCode: string;
	items: {
		id: string;
		name: string;
		price: number;
	}[];
	total: number;
}